feat: move GitHub client to separate package (#36)
